requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pyserial>=3.5
websockets>=12.0
python-multipart>=0.0.9
esptool>=4.7.0
